DSP pipeline warnings TODO fix.
DPOP #1 Warning DSP i_sq_reg multiplier stage i_sq_reg.P is not
pipelined (MREG=0). Pipelining the multiplier function will improve
performance and will save significant power so it is suggested whenever
possible to fully pipeline this function.  If this multiplier was
inferred, it is suggested to describe an additional register stage after
this function.  If there is no registered adder/accumulator following
the multiply function, two pipeline stages are suggested to allow both
the MREG and PREG registers to be used.  If the DSP48 was instantiated
in the design, it is suggested to set both the MREG and PREG attributes
to 1 when performing multiply functions.